+++ tests/testthat/test-data.R
@@ -86,3 +86,13 @@ test_that("CoL version info works", {
   
   expect_output(print(catalogue_of_life_version()))
 })
+
+test_that("CoNS/CoPS are up to date", {
+  uncategorised <- subset(microorganisms,
+                          genus == "Staphylococcus" &
+                            !species %in% c("", "aureus") &
+                            !mo %in% c(MO_CONS, MO_COPS))
+  expect(NROW(uncategorised) == 0,
+         failure_message = paste0("Staphylococcal species not categorised as CoNS/CoPS: S. ",
+                                  uncategorised$species, " (", uncategorised$mo, ")"))
+})